# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: Crawler.io API
  description: Unified Social Media Data API — one integration, all platforms.
  version: "1.0.0"
  contact:
    email: api@crawler.io
    url: https://docs.crawler.io
servers:
  - url: https://api.crawler.io/v1
    description: Production
  - url: https://sandbox.crawler.io/v1
    description: Sandbox (mock data)
security:
  - BearerAuth: []
...
paths:
...
  /webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - post.created
                      - post.updated
                      - metrics.updated
                      - creator.updated
                      - creator.new_post
                filters:
                  type: object
                  properties:
                    creator_ids:
                      type: array
                      items:
                        type: string
                    platforms:
                      type: array
                      items:
                        type: string
                secret:
                  type: string
                  description: Secret for HMAC signature verification
      responses:
        "201":
          description: Webhook registered
    get:
      tags: [Webhooks]
      summary: List registered webhooks
      responses:
        "200":
          description: Webhook list
  /webhooks/{webhook_id}:
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Webhook deleted

  # ─── TRENDS ─────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Pass your key as a Bearer token.
        Production keys: sk_live_*
        Sandbox keys: sk_test_*
...